refactor: Operator enum 분리
- 이제는 Operator 가 Caluclator에 종속적이지 않게 되었습니다.
  심지어 계산기는 Operator를 알지 못해도 상관 없는 상황이 되어 분리하였습니다.
